removing 3.5/pathlib.pyi
